modif size_t

+++ chamilo/pricer-skel/src/MonteCarlo.cpp
@@ -15,7 +15,7 @@ void MonteCarlo::price(double &prix, double &ic){
     prix = 0;
     PnlMat *path;
     double esp_carre; //premier membre pour calculer la variance
-    for (int j = 0; j < nbSamples_; ++j) {
+    for (size_t j = 0; j < nbSamples_; ++j) {
         mod_->asset(path, opt_->T_, opt_->nbTimeSteps_, rng_);
         prix += opt_->payoff(path);
         esp_carre += pow((opt_->payoff(path)),2);
@@ -23,7 +23,6 @@ void MonteCarlo::price(double &prix, double &ic){
     double estimateur = exp(-2*mod_->r_*opt_->T_)*((1/nbSamples_)*esp_carre-pow((1/nbSamples_)*prix,2));
     prix *= exp(-mod_->r_*opt_->T_)*(1/nbSamples_);
     ic = 2*1.96 * estimateur/sqrt(nbSamples_);
-    //ic = 2 * 1.96 * sqrt(estimateur/nbSamples_);
 }
 
 
@@ -31,7 +30,7 @@ void MonteCarlo::price(const PnlMat *past, double t, double &prix, double &ic){
     prix = 0;
     PnlMat *path;
     double esp_carre;
-    for (int i = 0; i < nbSamples_; ++i) {
+    for (size_t i = 0; i < nbSamples_; ++i) {
         mod_->asset(path, t, opt_->T_, opt_->nbTimeSteps_, rng_, past);
         prix += opt_->payoff(path);
         esp_carre += pow((opt_->payoff(path)),2);
@@ -56,7 +55,7 @@ void MonteCarlo::delta(const PnlMat *past, double t, PnlVect *delta, const PnlVe
     PnlMat *decrement_path;
 
     for (int d = 0; d < mod_->size_; d++) {
-      for (int i = 0; i < nbSamples_; i++) {
+      for (size_t i = 0; i < nbSamples_; i++) {
         mod_->asset(path, t, opt_->T_, opt_->nbTimeSteps_, rng_, past);
         mod_->shiftAsset(increment_path, path, d, fdStep_, t, timestep);
         mod_->shiftAsset(decrement_path, path, d, -fdStep_, t, timestep);
